Refactored duplicated code into handleJobProgress
